Move ExceptionWindow out of ui.py
